requirements.txt
# Backend API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0

# Data Augmentation
albumentations>=1.3.0

# Interpretability
captum>=0.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0
pydantic>=2.0.0

# Progress Bars
tqdm>=4.65.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

python-dotenv